:recycle: refactor: Refatorada implementação de Apólice

+++ src/main/java/br/dev/paulowolfgang/gestao_apolices/service/impl/ApoliceServiceImpl.java
@@ -1,5 +1,8 @@
 package br.dev.paulowolfgang.gestao_apolices.service.impl;
 
+import br.dev.paulowolfgang.gestao_apolices.dto.mapper.ApoliceMapper;
+import br.dev.paulowolfgang.gestao_apolices.dto.request.ApoliceRequestDto;
+import br.dev.paulowolfgang.gestao_apolices.dto.response.ApoliceResponseDto;
 import br.dev.paulowolfgang.gestao_apolices.entity.Apolice;
 import br.dev.paulowolfgang.gestao_apolices.repository.IApoliceRepository;
 import br.dev.paulowolfgang.gestao_apolices.service.IApoliceService;
@@ -18,48 +21,48 @@ public class ApoliceServiceImpl implements IApoliceService {
     }
 
     @Override
-    public Apolice salvar(Apolice apolice) {
-        return apoliceRepository.save(apolice);
+    public ApoliceResponseDto salvar(ApoliceRequestDto request) {
+        Apolice apolice = ApoliceMapper.converter(request);
+        apolice = apoliceRepository.save(apolice);
+        return ApoliceMapper.converter(apolice);
     }
 
     @Override
-    public Optional<Apolice> buscarPorId(Long id) {
-        return apoliceRepository.findById(id);
-    }
-
-    @Override
-    public List<Apolice> listarTodas() {
-        return apoliceRepository.findAll();
-    }
-
-    @Override
-    public Optional<Apolice> buscarPorNumero(String numero) {
-        return apoliceRepository.findByNumero(numero);
-    }
-
-    @Override
-    public Apolice atualizar(Long id, Apolice apoliceAtualizada) {
-
-        return apoliceRepository.findById(id)
-                .map(apoliceExistente -> {
-                    apoliceExistente.setCliente(apoliceAtualizada.getCliente());
-                    apoliceExistente.setNumero(apoliceAtualizada.getNumero());
-                    apoliceExistente.setValorCobertura(apoliceAtualizada.getValorCobertura());
-                    apoliceExistente.setPremioMensal(apoliceAtualizada.getPremioMensal());
-                    apoliceExistente.setPremioTotal(apoliceAtualizada.getPremioTotal());
-                    apoliceExistente.setParcelasTotais(apoliceAtualizada.getParcelasTotais());
-                    apoliceExistente.setParcelasPagas(apoliceAtualizada.getParcelasPagas());
-                    apoliceExistente.setDataInicio(apoliceAtualizada.getDataInicio());
-                    apoliceExistente.setDataFim(apoliceAtualizada.getDataFim());
-                    apoliceExistente.setTipo(apoliceAtualizada.getTipo());
-                    apoliceExistente.setStatus(apoliceAtualizada.getStatus());
-                    return apoliceRepository.save(apoliceExistente);
-                })
+    public ApoliceResponseDto buscarPorId(Long id) {
+        Apolice apolice = apoliceRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Apólice não encontrada para o ID: " + id));
+        return ApoliceMapper.converter(apolice);
+    }
+
+    @Override
+    public List<ApoliceResponseDto> listarTodos() {
+        List<Apolice> apolices = apoliceRepository.findAll();
+        return apolices.stream()
+                .map(ApoliceMapper::converter)
+                .toList();
+    }
+
+    @Override
+    public ApoliceResponseDto buscarPorNumero(String numero) {
+        Apolice apolice = apoliceRepository.findByNumero(numero)
+                .orElseThrow(() -> new IllegalArgumentException("Apólice não encontrada para o número: " + numero));
+        return ApoliceMapper.converter(apolice);
+    }
+
+    @Override
+    public ApoliceResponseDto atualizar(Long id, ApoliceRequestDto apoliceAtualizada) {
+        Apolice apolice = apoliceRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Apólice não encontrada para o ID: " + id));
+        ApoliceMapper.copiarParaPropriedades(apoliceAtualizada, apolice);
+        apolice = apoliceRepository.save(apolice);
+        return ApoliceMapper.converter(apolice);
     }
 
     @Override
     public void remover(Long id) {
+        if(!apoliceRepository.existsById(id)){
+            throw new IllegalArgumentException("Apólice não encontrada para o ID: " + id);
+        }
         apoliceRepository.deleteById(id);
     }
 }
